Move ProgressStatus to ui.rs
