+++ src/lib.rs
@@ -10,4 +10,5 @@ pub mod log;
 pub mod repo;
 pub mod repo_cache_serde;
 pub mod repo_name;
+pub mod ui;
 pub mod util;
